daleq-core/src/main/java/de/brands4friends/daleq/core/internal/formatting/DateFormatter.java
/*
 * Copyright 2012 brands4friends, Private Sale GmbH
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package de.brands4friends.daleq.core.internal.formatting;

import java.util.Date;

import org.joda.time.DateTime;
import org.joda.time.LocalDate;
import org.joda.time.LocalTime;
import org.joda.time.format.DateTimeFormatter;
import org.joda.time.format.DateTimeFormatterBuilder;

public final class DateFormatter {

    private static final DateTimeFormatter DATE_TIME_FORMATTER = new DateTimeFormatterBuilder()
            .appendPattern("yyyy-MM-dd HH:mm:ss.SSS")
            .toFormatter();

    private static final DateTimeFormatter LOCAL_DATE_FORMATTER = new DateTimeFormatterBuilder()
            .appendPattern("yyyy-MM-dd")
            .toFormatter();

    private static final DateTimeFormatter LOCAL_TIME_FORMATTER = new DateTimeFormatterBuilder()
            .appendPattern("HH:mm:ss")
            .toFormatter();


    private DateFormatter() {

    }


    public static String print(final DateTime dateTime) {
        return DATE_TIME_FORMATTER.print(dateTime);
    }

    public static String print(final Date date) {
        final DateTime dateTime = new DateTime(date);
        return print(dateTime);
    }

    public static String print(final LocalDate localDate) {
        return LOCAL_DATE_FORMATTER.print(localDate);
    }

    public static String print(final LocalTime localTime) {
        return LOCAL_TIME_FORMATTER.print(localTime);
    }
}
